requirements.txt
Flask
Flask-SocketIO
eventlet
pytube
requests
boto3
google-api-python-client
celery
python-dotenv
redis
certifi